Combining test cases and format.
